set tree

+++ pallets/template/src/types.rs
@@ -45,8 +45,8 @@ pub struct StakeDetails<Balance> {
 #[cfg_attr(feature = "std", derive(Debug))]
 pub struct SortitionSumTree {
     pub k: u64,
-    pub stack: Vec<u128>,
-    pub nodes: Vec<u128>,
+    pub stack: Vec<u64>,
+    pub nodes: Vec<u64>,
     pub ids_to_node_indexes: BTreeMap<u128, u64>, // citizen id, node index
     pub node_indexes_to_ids: BTreeMap<u64, u128>, // node index, citizen id
 }
